+++ alt/options.go
@@ -15,6 +15,16 @@ type Options struct {
 
 	// OmitNil if true omits object members that have nil values.
 	OmitNil bool
+
+	// UseTags if true will use the json annotation tags when marhsalling,
+	// writing, or decomposing an struct. If no tag is present then the
+	// KeyExact flag is referenced to determine the key.
+	UseTags bool
+
+	// KeyExact if true will use the exact field name for an encoded struct
+	// field. If false the key style most often seen in JSON files where the
+	// first character of the object keys is lowercase.
+	KeyExact bool
 }
 
 // DefaultOptions are the default options for decompsing.
@@ -22,4 +32,5 @@ var DefaultOptions = Options{
 	CreateKey:    "type",
 	FullTypePath: false,
 	OmitNil:      true,
+	UseTags:      false,
 }