+++ oj/write.go
@@ -50,6 +50,41 @@ func JSON(data interface{}, args ...interface{}) string {
 	return string(o.buf)
 }
 
+// Marshal returns a JSON string for the data provided. The data can be a
+// simple type of nil, bool, int, floats, time.Time, []interface{}, or
+// map[string]interface{} or a Node type, The args, if supplied can be an int
+// as an indent or a *Options. An error will be returned if the Option.Strict
+// flag is true and a value is encountered that can not be encoded other than
+// by using the %v format of the fmt package.
+func Marshal(data interface{}, args ...interface{}) ([]byte, error) {
+	o := &DefaultOptions
+	o.KeyExact = true
+	o.UseTags = true
+
+	if 0 < len(args) {
+		switch ta := args[0].(type) {
+		case int:
+			oi := *o
+			oi.Indent = ta
+			o = &oi
+		case *Options:
+			o = ta
+		}
+	}
+	o.strict = true
+	if o.InitSize == 0 {
+		o.InitSize = 256
+	}
+	if cap(o.buf) < o.InitSize {
+		o.buf = make([]byte, 0, o.InitSize)
+	} else {
+		o.buf = o.buf[:0]
+	}
+	err := o.buildJSON(data, 0)
+
+	return o.buf, err
+}
+
 // Write a JSON string for the data provided. The data can be a simple type of
 // nil, bool, int, floats, time.Time, []interface{}, or map[string]interface{}
 // or a Node type, The args, if supplied can be an int as an indent or a
@@ -172,13 +207,34 @@ func (o *Options) buildJSON(data interface{}, depth int) (err error) {
 			return o.buildJSON(data, depth)
 		}
 		if 0 < len(o.CreateKey) {
-			ao := alt.Options{CreateKey: o.CreateKey, OmitNil: o.OmitNil, FullTypePath: o.FullTypePath}
+			ao := alt.Options{
+				CreateKey:    o.CreateKey,
+				OmitNil:      o.OmitNil,
+				FullTypePath: o.FullTypePath,
+				UseTags:      o.UseTags,
+				KeyExact:     o.KeyExact,
+			}
 			return o.buildJSON(alt.Decompose(data, &ao), depth)
+		}
+		if !o.NoReflect {
+			ao := alt.Options{
+				CreateKey:    o.CreateKey,
+				OmitNil:      o.OmitNil,
+				FullTypePath: o.FullTypePath,
+				UseTags:      o.UseTags,
+				KeyExact:     o.KeyExact,
+			}
+			if dec := alt.Decompose(data, &ao); dec != nil {
+				return o.buildJSON(dec, depth)
+			}
+		}
+		if o.strict {
+			err = fmt.Errorf("%T can not be encoded as a JSON element", data)
 		} else {
 			o.buildString(fmt.Sprintf("%v", td))
 		}
 	}
-	if o.w != nil && o.WriteLimit < len(o.buf) {
+	if err == nil && o.w != nil && o.WriteLimit < len(o.buf) {
 		_, err = o.w.Write(o.buf)
 		o.buf = o.buf[:0]
 	}